% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fraction}
\alias{fraction}
\title{Gap fraction extractor}
\usage{
fraction(mat, Zenith = 18, Azimut = 24, declination = 0)
}
\arguments{
\item{mat}{a matrix with 0s and 1s}

\item{Zenith}{integer; the number of divisions in the vertical plane}

\item{Azimut}{integer; the number of sectors in the horizontal plant}

\item{declination}{the magnetic declination in degrees at the site}
}
\value{
a list
}
\description{
Gap fraction extractor (white pixels) for each sky segment which are defined by number of zenith annuli (vertical plane) and azimuth sectors (horizontal plane).
}
\details{
The number of segments are controlled by the \code{Zenith}  and \code{Azimut} arguments. A recomended division is \code{Zenith}=18 and \code{Azimut}=24 which are the default settings.
}
\examples{
# Define three points on the fish-eye ring
pointA <- c(x = 140, y = 102)
pointB <- c(x = 401.5, y = 25.9)
pointC <- c(x = 506.6, y = 238.6)
DFpoints <- data.frame(A = pointA, B = pointB, C = pointC)
# Read image and convert to black and white
blue <- ReadImg(system.file("pictures", "fisheye.jpg", package="cimesr"),
                channel = 3, coord = DFpoints, crop = TRUE)
bw <- threshold(blue, thresValue = 0.73)
# Extract gap fractions
frac <- fraction(bw, Zenith = 18, Azimut = 24, declination = 0)
}
\author{
Gustaf Granath, Alemu Gonsamo, Jean-Michel Walter
}
\seealso{
gap size extraction will be added
}
\keyword{extraction}
\keyword{gap}

