% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReadImg}
\alias{ReadImg}
\title{Read jpg/bmp image}
\usage{
ReadImg(filename, channel = 3, coord = NULL, crop = TRUE)
}
\arguments{
\item{filename}{character; a jpg or bmp file}

\item{channel}{integer (1-3); which RGB channel to store, 3 = blue.}

\item{coord}{data frame or matrix with three pixel coordinates. Three columns where row one has x coordinates and row two y coordinates. See example.}

\item{crop}{logical; if TRUE, the hemispherical image is cropped out.}
}
\value{
a matrix with cell values between 0 and 1.
}
\description{
Reads a hemispherical (fish-eye) jpg/bmp file. Returns a specified color channel and the image area is cropped out.
}
\details{
This function calls the \code{read.bitmap()} function in the \code{readbitmap} package. The recomended channel to use for canopy analyses is the blue channel. This function crop the image so a qudratic image of the fish-eye area is returned. Although cropping is needed for any gap analyses in cimesr, it can be disabled by setting \code{crop = FALSE}. The whole image is then returned. The output is always a matrix with cell values from 0 to 1 and represents a greyscale.
}
\examples{
# Define three points on the fish-eye ring
pointA <- c(x = 128.4, y = 92.9)
pointB <- c(x = 381.7, y = 17.3)
pointC <- c(x = 494.9, y = 225.4)
DFpoints <- data.frame(A = pointA, B = pointB, C = pointC)
blue <- ReadImg(system.file("pictures", "fisheye.jpg", package="cimesr"),
                channel = 3, coord = DFpoints, crop = TRUE)
}
\author{
Gustaf Granath
}
\keyword{image}
\keyword{read}

