% Generated by roxygen2 (4.0.2): do not edit by hand
\name{size}
\alias{size}
\title{Gap size extractor}
\usage{
size(mat, MinZenith = 0, MaxZenith = 90, step = 1)
}
\arguments{
\item{mat}{a matrix with 0s and 1s}

\item{MinZenith}{numeric; minimum zenith angle in degrees (0 to 90)}

\item{MaxZenith}{numeric;  maximum zenith angle in degrees (0 to 90)}

\item{step}{numeric; angle step (1 degree is recomended)}
}
\value{
a list

\code{ZenithWhite} zenith angle, degrees, corresponding to the white sequence

\code{WhiteSeq} lengths (in pixel units) of sequences of white pixels

\code{ZenithBlack} zenith angle, degrees, corresponding to the black sequence

\code{BlackSeq} lengths (in pixel units) of sequences of black pixels
}
\description{
Extract gap sizes along 'circular transects' at a range of zenith angles.
}
\details{
Sequences of black and white pixels start at the first edge from gap to non-gap or vice-
versa, for each zenith angle. This is called foliage clumping. Black pixels
represent foliage or sky obstruction, whereas white pixels represent sky.
}
\examples{
# Define three points on the fish-eye ring
pointA <- c(x = 140, y = 102)
pointB <- c(x = 401.5, y = 25.9)
pointC <- c(x = 506.6, y = 238.6)
DFpoints <- data.frame(A = pointA, B = pointB, C = pointC)
# Read image and convert to black and white
blue <- ReadImg(system.file("pictures", "fisheye.jpg", package="cimesr"),
                channel = 3, coord = DFpoints, crop = TRUE)
bw <- threshold(blue, thresValue = 0.73)
# Extract gap size
siz <- size(bw,  MinZenith = 0, MaxZenith = 90, step = 1)
}
\author{
Gustaf Granath, Alemu Gonsamo, Jean-Michel Walter
}
\keyword{extraction}
\keyword{gap}

