% Generated by roxygen2 (4.0.2): do not edit by hand
\name{threshold}
\alias{threshold}
\title{Greyscale to black and white}
\usage{
threshold(mat, thresValue)
}
\arguments{
\item{mat}{numeric matrix}

\item{thresValue}{numeric; a value defining the threshold between white and black. Must be between max and min value of /code{mat}.}
}
\value{
matrix with cell values 0 (white) or 1 (black).
}
\description{
Convert a greyscale image (matrix) into black and white using a set threshold.
}
\details{
A threshold value can be estimated from imaging software or by using the /code{AutoThres()} function.
}
\examples{
# Read image and convert to black and white
blue <- readbitmap::read.bitmap(system.file("pictures", "fisheye.jpg",
                                package="cimesr"), channel = 3)
bw <- threshold(blue, thresValue = 0.73)
}
\author{
Gustaf Granath, Alemu Gonsamo, Jean-Michel Walter
}

